댓글 추천 비추천 로직 수정

+++ client-react/src/common/pages/post/unknown/UnknownReply.js
@@ -191,35 +191,37 @@ const UnknownReply = (props) => {
 					);
 
 					for (const replyData of replyArray[0]) {
+						let isDeleted = (replyData.deletedAt !== null);
+						let deleteButton = <></>;
+
+						if(isDeleted === false){
+							deleteButton = (
+								<div onClick={() => {deleteReply(replyData.code, currentPage)}} style={{ display: "flex", alignItems: "center",cursor: "pointer" }}>
+									<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="palevioletred" className="bi bi-x-circle" viewBox="0 0 16 16">
+										<path d="M8 15A7 7 0 1 1 8 1a7 7 0 0 1 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z"/>
+										<path d="M4.646 4.646a.5.5 0 0 1 .708 0L8 7.293l2.646-2.647a.5.5 0 0 1 .708.708L8.707 8l2.647 2.646a.5.5 0 0 1-.708.708L8 8.707l-2.646 2.647a.5.5 0 0 1-.708-.708L7.293 8 4.646 5.354a.5.5 0 0 1 0-.708z"/>
+									</svg>
+								</div>
+							);
+						}
+
 						const replyBody = (
 							<div style={{width: "100%", display: "flex", flexDirection: "column", paddingBottom: "5px", marginBottom: "5px"}}>
 								<div style={{display: "flex", justifyContent: "space-between", alignItems: "center"}}>
 									<div>
 										<span style={{fontSize: "0.8rem", color: "gray"}}>익명 ({replyData.ip})</span>
 									</div>
-									
-
-									<div onClick={() => {deleteReply(replyData.code, currentPage)}} style={{ display: "flex", alignItems: "center",cursor: "pointer" }}>
-										{
-											replyData.deletedAt === null ?
-											<>
-												<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="palevioletred" className="bi bi-x-circle" viewBox="0 0 16 16">
-													<path d="M8 15A7 7 0 1 1 8 1a7 7 0 0 1 0 14zm0 1A8 8 0 1 0 8 0a8 8 0 0 0 0 16z"/>
-													<path d="M4.646 4.646a.5.5 0 0 1 .708 0L8 7.293l2.646-2.647a.5.5 0 0 1 .708.708L8.707 8l2.647 2.646a.5.5 0 0 1-.708.708L8 8.707l-2.646 2.647a.5.5 0 0 1-.708-.708L7.293 8 4.646 5.354a.5.5 0 0 1 0-.708z"/>
-												</svg>
-											</>
-											:
-											<></>
-										}
-									</div>
+									{deleteButton}
 								</div>
 
 								<div style={{fontSize: "0.75rem", marginTop: "5px", whiteSpace: "break-spaces", overflowWrap: "anywhere", overflow: "auto", marginRight: "3%"}}>
-									{replyData.deletedAt === null ? replyData.content : <span style={{color: "palevioletred"}}>{replyData.content}</span>}
+									<span style={{color: isDeleted === false ? "black" : "palevioletred"}}>
+										{replyData.content}
+									</span>
 								</div>
 
 								<p style={{fontSize: "0.75rem", color: "lightgray", marginTop: "0.8rem"}}>
-									{new Date(replyData.createdAt).toLocaleString("sv-SE")}
+									{new Date(replyData.createdAt).toLocaleString("sv-SE").substring(0, 16)}
 								</p>
 							</div>
 						);
